+++ nodebackend/src/routes/simpleapiRoute.ts
@@ -1,15 +1,15 @@
-const express = require('express');
-const router = express.Router();
-const MqttPublisher = require('../../nodebackend/build/utils/mqttPublisher').default;
-const logger = require('../../nodebackend/build/logger').default;
+import express from 'express';
+import MqttPublisher from '../utils/mqttPublisher';
+import logger from '../logger';
 
+const router = express.Router();
 const publisher = new MqttPublisher();
 
-router.post('/', async function (req, res) {
+router.post('/', async (req: express.Request, res: express.Response) => {
     const { topic, state } = req.body;
 
     try {
-        publisher.publish(topic, state.toString(), (err) => {
+        publisher.publish(topic, state.toString(), (err: Error | null) => {
             if (err) {
                 logger.error('Error while publishing message:', err);
                 res.status(500).send('Error while publishing message to MQTT broker.');
@@ -23,4 +23,4 @@ router.post('/', async function (req, res) {
     }
 });
 
-module.exports = router;
+export default router;