+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] 2023-12-16
+### Changed
+- Migration of simpleapiRoute to TS
+
 ## [v15.4.0] 2023-12-10
 ### Changed
 - Migration of mongoClient, buildUrlMap and mqttPublisher to TS